+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/state/ControlPanelState.kt
@@ -97,7 +97,7 @@ data class ControlPanelState(
         val type: Type
     ) {
         enum class Type {
-            P, H1, H2, H3, H4, TITLE, QUOTE, CODE_SNIPPET, BULLET, NUMBERED, TOGGLE, CHECKBOX
+            P, H1, H2, H3, H4, TITLE, QUOTE, CODE_SNIPPET, BULLET, NUMBERED, TOGGLE, CHECKBOX, BOOKMARK
         }
     }